scan_full_port/requirements.txt
psycopg2-binary
python-nmap
confluent-kafka
dotenv
schedule